nvim: Add Neomake plugin and corresponding autocmd
Neomake (https://github.com/neomake/neomake) is an "[a]synchronous
linting and make framework for Neovim/Vim".  We want it to run after
every buffer write.  This is accomplished with an autocmd on
BufWritePost.

The Neomake autocmd is in an augroup to protect it from being sourced
twice.  That augroup in turn is in a file inside the plugin directory;
the contents of plugin are recursively sourced by nvim instances as part
of startup.